refactor: extract backup feature folder
Move backup-* modules from src/lib/ and src/components/ into
src/app/features/backup/{lib,widgets} with a public index.ts barrel.
First step of the feature-folder migration.
