+++ src/app/features/backup/index.ts
@@ -1,0 +1,2 @@
+export { BackupSubscriber, SpacesBackupSubscriber } from "./widgets/subscribers"
+export { BackupSettings, SpaceBackupSettings } from "./widgets/settings-panel"
