+++ src/app/main.tsx
@@ -11,10 +11,7 @@ import {
 	useSplashDelay,
 } from "@/components/splash-screen"
 import { PWAContext, usePWAProvider, PWAInstallHint } from "@/lib/pwa"
-import {
-	BackupSubscriber,
-	SpacesBackupSubscriber,
-} from "@/lib/backup-subscribers"
+import { BackupSubscriber, SpacesBackupSubscriber } from "@/app/features/backup"
 import { useCleanupDeleted } from "@/lib/use-cleanup-deleted"
 import { init } from "@plausible-analytics/tracker"
